// old/src/lib/types/product.ts
export type Product = {
    _id: string
    title: string
    price: number
    category: {
        id: string
        name: string
    }
    rating: number
    tags?: string[]
    thumbnail: {
        url: string
    },
    new: boolean
    sale: boolean
    hot: boolean
    trending: boolean
    images: [{ url: string }]
    discountType: string
    discountPrice: number
    stock: number
    slug: string
    isVarientStatus: boolean
    meta?: {
        keywords: string
        meta_description: string
    };
    reviews?: number;
    short_description?: string;
    quantity?: number;
    sku?: string;

}